refactor(router) move path construction to after_access
The router was repsonsible for path construction but this would
happen in the before-access phase. Allowing path manipulation to
only be done on the full path, not on the individual components.

This commit changes it by creating an intermediate table with the
path components (for plugins to modify) and does the final
construction only in the after-access phase
