+++ kong/router.lua
@@ -1185,35 +1185,15 @@ function _M.new(routes)
           end
 
           if matched_route then
+            local request_postfix
             local upstream_host
-            local upstream_uri
             local upstream_url_t = matched_route.upstream_url_t
             local matches        = ctx.matches
-
-            -- Path construction
 
             if matched_route.type == "http" then
               -- if we do not have a path-match, then the postfix is simply the
               -- incoming path, without the initial slash
-              local request_postfix = matches.uri_postfix or sub(req_uri, 2, -1)
-              local upstream_base = upstream_url_t.path or "/"
-
-              if matched_route.strip_uri then
-                -- we drop the matched part, replacing it with the upstream path
-                if sub(upstream_base, -1, -1) == "/" and
-                   sub(request_postfix, 1, 1) == "/" then
-                  -- double "/", so drop the first
-                  upstream_uri = sub(upstream_base, 1, -2) .. request_postfix
-
-                else
-                  upstream_uri = upstream_base .. request_postfix
-                end
-
-              else
-                -- we retain the incoming path, just prefix it with the upstream
-                -- path, but skip the initial slash
-                upstream_uri = upstream_base .. sub(req_uri, 2, -1)
-              end
+              request_postfix = matches.uri_postfix or sub(req_uri, 2, -1)
 
               -- preserve_host header logic
 
@@ -1222,24 +1202,24 @@ function _M.new(routes)
               end
             end
 
-            local match_t     = {
-              route           = matched_route.route,
-              service         = matched_route.service,
-              headers         = matched_route.headers,
-              upstream_url_t  = upstream_url_t,
-              upstream_scheme = upstream_url_t.scheme,
-              upstream_uri    = upstream_uri,
-              upstream_host   = upstream_host,
-              matches         = {
-                uri_captures  = matches.uri_captures,
-                uri           = matches.uri,
-                host          = matches.host,
-                method        = matches.method,
-                src_ip        = matches.src_ip,
-                src_port      = matches.src_port,
-                dst_ip        = matches.dst_ip,
-                dst_port      = matches.dst_port,
-                sni           = matches.sni,
+            local match_t = {
+              route                = matched_route.route,
+              service              = matched_route.service,
+              headers              = matched_route.headers,
+              upstream_url_t       = upstream_url_t,
+              upstream_scheme      = upstream_url_t.scheme,
+              upstream_host        = upstream_host,
+              upstream_uri_postfix = request_postfix,
+              matches = {
+                uri_captures       = matches.uri_captures,
+                uri                = matches.uri,
+                host               = matches.host,
+                method             = matches.method,
+                src_ip             = matches.src_ip,
+                src_port           = matches.src_port,
+                dst_ip             = matches.dst_ip,
+                dst_port           = matches.dst_port,
+                sni                = matches.sni,
               }
             }
