+++ spec/01-unit/08-router_spec.lua
@@ -1207,7 +1207,7 @@ describe("Router", function()
 
       -- upstream_uri
       assert.is_nil(match_t.upstream_host) -- only when `preserve_host = true`
-      assert.equal("/my-route", match_t.upstream_uri)
+      assert.equal("", match_t.upstream_uri_postfix)
 
       _ngx = mock_ngx("GET", "/my-route-2", { host = "domain.org" })
       router._set_ngx(_ngx)
@@ -1221,7 +1221,7 @@ describe("Router", function()
 
       -- upstream_uri
       assert.is_nil(match_t.upstream_host) -- only when `preserve_host = true`
-      assert.equal("/my-route-2", match_t.upstream_uri)
+      assert.equal("", match_t.upstream_uri_postfix)
     end)
 
     it("returns matched_host + matched_uri + matched_method", function()
@@ -1371,7 +1371,7 @@ describe("Router", function()
       local _ngx = mock_ngx("GET", "/hello/world", { host = "domain.org" })
       router._set_ngx(_ngx)
       local match_t = router.exec()
-      assert.equal("/world", match_t.upstream_uri)
+      assert.equal("/world", match_t.upstream_uri_postfix)
       assert.is_nil(match_t.matches.uri_captures)
     end)
 
@@ -1469,7 +1469,7 @@ describe("Router", function()
       router._set_ngx(_ngx)
       local match_t = router.exec()
       assert.same(use_case_routes[1].route, match_t.route)
-      assert.equal("/endel%C3%B8st", match_t.upstream_uri)
+      assert.equal("", match_t.upstream_uri_postfix)
     end)
 
     describe("stripped paths", function()
@@ -1502,7 +1502,7 @@ describe("Router", function()
         router._set_ngx(_ngx)
         local match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/hello/world", match_t.upstream_uri)
+        assert.equal("/hello/world", match_t.upstream_uri_postfix)
       end)
 
       it("strips if matched URI is plain (not a prefix)", function()
@@ -1510,7 +1510,7 @@ describe("Router", function()
         router._set_ngx(_ngx)
         local match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
       end)
 
       it("doesn't strip if 'strip_uri' is not enabled", function()
@@ -1519,7 +1519,7 @@ describe("Router", function()
         router._set_ngx(_ngx)
         local match_t = router.exec()
         assert.same(use_case_routes[2].route, match_t.route)
-        assert.equal("/my-route/hello/world", match_t.upstream_uri)
+        assert.equal("/hello/world", match_t.upstream_uri_postfix)
       end)
 
       it("does not strips root / URI", function()
@@ -1540,7 +1540,7 @@ describe("Router", function()
         router._set_ngx(_ngx)
         local match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/my-route/hello/world", match_t.upstream_uri)
+        assert.equal("my-route/hello/world", match_t.upstream_uri_postfix)
       end)
 
       it("can find an route with stripped URI several times in a row", function()
@@ -1548,13 +1548,13 @@ describe("Router", function()
         router._set_ngx(_ngx)
         local match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
 
         _ngx = mock_ngx("GET", "/my-route", { host = "domain.org" })
         router._set_ngx(_ngx)
         match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
       end)
 
       it("can proxy an route with stripped URI with different URIs in a row", function()
@@ -1562,25 +1562,25 @@ describe("Router", function()
         router._set_ngx(_ngx)
         local match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
 
         _ngx = mock_ngx("GET", "/this-route", { host = "domain.org" })
         router._set_ngx(_ngx)
         match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
 
         _ngx = mock_ngx("GET", "/my-route", { host = "domain.org" })
         router._set_ngx(_ngx)
         match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
 
         _ngx = mock_ngx("GET", "/this-route", { host = "domain.org" })
         router._set_ngx(_ngx)
         match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
       end)
 
       it("strips url encoded paths", function()
@@ -1599,7 +1599,7 @@ describe("Router", function()
         router._set_ngx(_ngx)
         local match_t = router.exec()
         assert.same(use_case_routes[1].route, match_t.route)
-        assert.equal("/", match_t.upstream_uri)
+        assert.equal("", match_t.upstream_uri_postfix)
       end)
 
       it("strips a [uri regex]", function()
@@ -1618,7 +1618,7 @@ describe("Router", function()
                               { host = "domain.org" })
         router._set_ngx(_ngx)
         local match_t = router.exec()
-        assert.equal("/hello/world", match_t.upstream_uri)
+        assert.equal("/hello/world", match_t.upstream_uri_postfix)
       end)
 
       it("strips a [uri regex] with a capture group", function()
@@ -1637,7 +1637,7 @@ describe("Router", function()
                               { host = "domain.org" })
         router._set_ngx(_ngx)
         local match_t = router.exec()
-        assert.equal("/hello/world", match_t.upstream_uri)
+        assert.equal("/hello/world", match_t.upstream_uri_postfix)
       end)
     end)
 
@@ -1778,193 +1778,6 @@ describe("Router", function()
           assert.is_nil(match_t.upstream_host)
         end)
       end)
-    end)
-
-
-    describe("slash handling", function()
-      local checks = {
-        -- upstream url    paths           request path    expected path           strip uri
-        {  "/",            "/",            "/",            "/",                    true      }, -- 1
-        {  "/",            "/",            "/foo/bar",     "/foo/bar",             true      },
-        {  "/",            "/",            "/foo/bar/",    "/foo/bar/",            true      },
-        {  "/",            "/foo/bar",     "/foo/bar",     "/",                    true      },
-        {  "/",            "/foo/bar",     "/foo/bar/",    "/",                    true      },
-        {  "/",            "/foo/bar/",    "/foo/bar/",    "/",                    true      },
-        {  "/fee/bor",     "/",            "/",            "/fee/bor",             true      },
-        {  "/fee/bor",     "/",            "/foo/bar",     "/fee/borfoo/bar",      true      },
-        {  "/fee/bor",     "/",            "/foo/bar/",    "/fee/borfoo/bar/",     true      },
-        {  "/fee/bor",     "/foo/bar",     "/foo/bar",     "/fee/bor",             true      }, -- 10
-        {  "/fee/bor",     "/foo/bar",     "/foo/bar/",    "/fee/bor/",            true      },
-        {  "/fee/bor",     "/foo/bar/",    "/foo/bar/",    "/fee/bor",             true      },
-        {  "/fee/bor/",    "/",            "/",            "/fee/bor/",            true      },
-        {  "/fee/bor/",    "/",            "/foo/bar",     "/fee/bor/foo/bar",     true      },
-        {  "/fee/bor/",    "/",            "/foo/bar/",    "/fee/bor/foo/bar/",    true      },
-        {  "/fee/bor/",    "/foo/bar",     "/foo/bar",     "/fee/bor/",            true      },
-        {  "/fee/bor/",    "/foo/bar",     "/foo/bar/",    "/fee/bor/",            true      },
-        {  "/fee/bor/",    "/foo/bar/",    "/foo/bar/",    "/fee/bor/",            true      },
-        {  "/",            "/",            "/",            "/",                    false     },
-        {  "/",            "/",            "/foo/bar",     "/foo/bar",             false     }, -- 20
-        {  "/",            "/",            "/foo/bar/",    "/foo/bar/",            false     },
-        {  "/",            "/foo/bar",     "/foo/bar",     "/foo/bar",             false     },
-        {  "/",            "/foo/bar",     "/foo/bar/",    "/foo/bar/",            false     },
-        {  "/",            "/foo/bar/",    "/foo/bar/",    "/foo/bar/",            false     },
-        {  "/fee/bor",     "/",            "/",            "/fee/bor",             false     },
-        {  "/fee/bor",     "/",            "/foo/bar",     "/fee/borfoo/bar",      false     },
-        {  "/fee/bor",     "/",            "/foo/bar/",    "/fee/borfoo/bar/",     false     },
-        {  "/fee/bor",     "/foo/bar",     "/foo/bar",     "/fee/borfoo/bar",      false     },
-        {  "/fee/bor",     "/foo/bar",     "/foo/bar/",    "/fee/borfoo/bar/",     false     },
-        {  "/fee/bor",     "/foo/bar/",    "/foo/bar/",    "/fee/borfoo/bar/",     false     }, -- 30
-        {  "/fee/bor/",    "/",            "/",            "/fee/bor/",            false     },
-        {  "/fee/bor/",    "/",            "/foo/bar",     "/fee/bor/foo/bar",     false     },
-        {  "/fee/bor/",    "/",            "/foo/bar/",    "/fee/bor/foo/bar/",    false     },
-        {  "/fee/bor/",    "/foo/bar",     "/foo/bar",     "/fee/bor/foo/bar",     false     },
-        {  "/fee/bor/",    "/foo/bar",     "/foo/bar/",    "/fee/bor/foo/bar/",    false     },
-        {  "/fee/bor/",    "/foo/bar/",    "/foo/bar/",    "/fee/bor/foo/bar/",    false     },
-        -- the following block runs the same tests, but with a request path that is longer
-        -- than the matched part, so either matches in the middle of a segment, or has an
-        -- additional segment.
-        {  "/",            "/",            "/foo/bars",    "/foo/bars",            true      },
-        {  "/",            "/",            "/foo/bar/s",   "/foo/bar/s",           true      },
-        {  "/",            "/foo/bar",     "/foo/bars",    "/s",                   true      },
-        {  "/",            "/foo/bar/",    "/foo/bar/s",   "/s",                   true      }, -- 40
-        {  "/fee/bor",     "/",            "/foo/bars",    "/fee/borfoo/bars",     true      },
-        {  "/fee/bor",     "/",            "/foo/bar/s",   "/fee/borfoo/bar/s",    true      },
-        {  "/fee/bor",     "/foo/bar",     "/foo/bars",    "/fee/bors",            true      },
-        {  "/fee/bor",     "/foo/bar/",    "/foo/bar/s",   "/fee/bors",            true      },
-        {  "/fee/bor/",    "/",            "/foo/bars",    "/fee/bor/foo/bars",    true      },
-        {  "/fee/bor/",    "/",            "/foo/bar/s",   "/fee/bor/foo/bar/s",   true      },
-        {  "/fee/bor/",    "/foo/bar",     "/foo/bars",    "/fee/bor/s",           true      },
-        {  "/fee/bor/",    "/foo/bar/",    "/foo/bar/s",   "/fee/bor/s",           true      },
-        {  "/",            "/",            "/foo/bars",    "/foo/bars",            false     },
-        {  "/",            "/",            "/foo/bar/s",   "/foo/bar/s",           false     }, -- 50
-        {  "/",            "/foo/bar",     "/foo/bars",    "/foo/bars",            false     },
-        {  "/",            "/foo/bar/",    "/foo/bar/s",   "/foo/bar/s",           false     },
-        {  "/fee/bor",     "/",            "/foo/bars",    "/fee/borfoo/bars",     false     },
-        {  "/fee/bor",     "/",            "/foo/bar/s",   "/fee/borfoo/bar/s",    false     },
-        {  "/fee/bor",     "/foo/bar",     "/foo/bars",    "/fee/borfoo/bars",     false     },
-        {  "/fee/bor",     "/foo/bar/",    "/foo/bar/s",   "/fee/borfoo/bar/s",    false     },
-        {  "/fee/bor/",    "/",            "/foo/bars",    "/fee/bor/foo/bars",    false     },
-        {  "/fee/bor/",    "/",            "/foo/bar/s",   "/fee/bor/foo/bar/s",   false     },
-        {  "/fee/bor/",    "/foo/bar",     "/foo/bars",    "/fee/bor/foo/bars",    false     },
-        {  "/fee/bor/",    "/foo/bar/",    "/foo/bar/s",   "/fee/bor/foo/bar/s",   false     }, -- 60
-        -- the following block matches on host, instead of path
-        {  "/",            nil,            "/",            "/",                    false     },
-        {  "/",            nil,            "/foo/bar",     "/foo/bar",             false     },
-        {  "/",            nil,            "/foo/bar/",    "/foo/bar/",            false     },
-        {  "/fee/bor",     nil,            "/",            "/fee/bor",             false     },
-        {  "/fee/bor",     nil,            "/foo/bar",     "/fee/borfoo/bar",      false     },
-        {  "/fee/bor",     nil,            "/foo/bar/",    "/fee/borfoo/bar/",     false     },
-        {  "/fee/bor/",    nil,            "/",            "/fee/bor/",            false     },
-        {  "/fee/bor/",    nil,            "/foo/bar",     "/fee/bor/foo/bar",     false     },
-        {  "/fee/bor/",    nil,            "/foo/bar/",    "/fee/bor/foo/bar/",    false     },
-        {  "/",            nil,            "/",            "/",                    true      }, -- 70
-        {  "/",            nil,            "/foo/bar",     "/foo/bar",             true      },
-        {  "/",            nil,            "/foo/bar/",    "/foo/bar/",            true      },
-        {  "/fee/bor",     nil,            "/",            "/fee/bor",             true      },
-        {  "/fee/bor",     nil,            "/foo/bar",     "/fee/borfoo/bar",      true      },
-        {  "/fee/bor",     nil,            "/foo/bar/",    "/fee/borfoo/bar/",     true      },
-        {  "/fee/bor/",    nil,            "/",            "/fee/bor/",            true      },
-        {  "/fee/bor/",    nil,            "/foo/bar",     "/fee/bor/foo/bar",     true      },
-        {  "/fee/bor/",    nil,            "/foo/bar/",    "/fee/bor/foo/bar/",    true      },
-      }
-
-      for i, args in ipairs(checks) do
-
-        local config
-        if args[5] == true then
-          config = "(strip = on, plain)"
-        else
-          config = "(strip = off, plain)"
-        end
-
-        local description
-        if args[2] then
-          description = string.format("(%d) (%s) %s with uri %s when requesting %s",
-            i, config, args[1], args[2], args[3])
-        else
-          description = string.format("(%d) (%s) %s with host %s when requesting %s",
-            i, config, args[1], "localbin-" .. i .. ".com", args[3])
-        end
-
-        it(description, function()
-          local use_case_routes = {
-            {
-              service      = {
-                protocol   = "http",
-                name       = "service-invalid",
-                path       = args[1],
-              },
-              route        = {
-                strip_path = args[5],
-                paths      = { args[2] },
-              },
-              headers      = {
-                -- only add the header is no path is provided
-                host       = args[2] == nil and nil or { "localbin-" .. i .. ".com" },
-              },
-            }
-          }
-
-          local router = assert(Router.new(use_case_routes) )
-          local _ngx = mock_ngx("GET", args[3], { host = "localbin-" .. i .. ".com" })
-          router._set_ngx(_ngx)
-          local match_t = router.exec()
-          assert.same(use_case_routes[1].route, match_t.route)
-          assert.equal(args[1], match_t.upstream_url_t.path)
-          assert.equal(args[4], match_t.upstream_uri)
-        end)
-      end
-
-      -- this is identical to the tests above, except that for the path we match
-      -- with an injected regex sequence, effectively transforming the path
-      -- match into a regex match
-      local function make_a_regex(path)
-        return "/[0]?" .. path:sub(2, -1)
-      end
-
-      for i, args in ipairs(checks) do
-        local config
-        if args[5] == true then
-          config = "(strip = on, regex)"
-        else
-          config = "(strip = off, regex)"
-        end
-
-        if args[2] then -- skip test cases which match on host
-          local description = string.format("(%d) (%s) %s with uri %s when requesting %s",
-                                            i, config, args[1], make_a_regex(args[2]), args[3])
-
-          it(description, function()
-
-
-            local use_case_routes = {
-              {
-                service      = {
-                  protocol   = "http",
-                  name       = "service-invalid",
-                  path       = args[1],
-                },
-                route        = {
-                  strip_path = args[5],
-                  paths      = { make_a_regex(args[2]) },
-                },
-                headers      = {
-                  -- only add the header is no path is provided
-                  host       = args[2] == nil and nil or { "localbin-" .. i .. ".com" },
-                },
-              }
-            }
-
-            local router = assert(Router.new(use_case_routes) )
-            local _ngx = mock_ngx("GET", args[3], { host = "localbin-" .. i .. ".com" })
-            router._set_ngx(_ngx)
-            local match_t = router.exec()
-            assert.same(use_case_routes[1].route, match_t.route)
-            assert.equal(args[1], match_t.upstream_url_t.path)
-            assert.equal(args[4], match_t.upstream_uri)
-          end)
-        end
-      end
     end)
   end)
